requirements.txt
python-docx==0.8.6
appdirs==1.4.3
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.11
enum34==1.1.6
ipdb==0.10.2
ipython==5.3.0
ipython-genutils==0.2.0
lxml==4.5.1
packaging==16.8
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.2.0
python-docx==0.8.6
rsa==3.4.2
simplegeneric==0.8.1
traitlets==4.3.2
uritemplate==3.0.0
wcwidth==0.1.7
gdata==2.0.18
google-api-python-client>=1.9.1
httplib2==0.18.1
oauth2client==4.1.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
rsa==3.4.2
six==1.11.0
uritemplate==3.0.0
markdown==3.1.1
weasyprint==0.42.0
html2text==2020.1.16
